+++ src/net/mcmc_cfg.rs
@@ -8,8 +8,8 @@ use std::{
 
 /// Parameters for MCMC sampling.
 pub struct MCMCCfg {
-    pub hmc_step_size_factor: f64,
-    pub hmc_max_hamiltonian_error: f64,
+    pub hmc_step_size_factor: f32,
+    pub hmc_max_hamiltonian_error: f32,
     pub hmc_integration_length: usize,
     pub hmc_step_size_mode: StepSizeMode,
     pub chain_length: usize,
